+++ UI/Services/ComponentService.cs
@@ -45,9 +45,9 @@
             if (component == null) {
                 this.SelectedItem = null;
             }
-            else if (this._sceneService.CurrentScene is SceneWrapper scene) {
+            else if (this._sceneService.CurrentScene != null) {
                 ComponentWrapper result = null;
-                foreach (var child in scene.Children) {
+                foreach (var child in this._sceneService.CurrentScene.Children) {
                     if (child.Id == component.Id) {
                         result = child;
                     }